Refactor metrics file structure.
Separate metrics into individual files.
